🚚 core: Split Custom* into a separate mod, `custom`
